+++ spec/components/image_card_spec.rb
@@ -117,6 +117,12 @@ describe "ImageCard", type: :view do
     assert_select ".gem-c-image-card__image[height='90']"
   end
 
+  it "renders two thirds variant with large title and description text font size for mobile" do
+    render_component(href: "#", image_src: "/moo.jpg", image_alt: "some meaningful alt text", heading_text: "heading", description: "description", two_thirds: true, large_font_size_mobile: true)
+    assert_select ".gem-c-image-card--two-thirds .gem-c-image-card__title-link--large-font-size-mobile", text: "heading"
+    assert_select ".gem-c-image-card--two-thirds .gem-c-image-card__description--large-font-size-mobile", text: "description"
+  end
+
   it "applies tracking attributes" do
     render_component(href: "#", href_data_attributes: { track_category: "cat" }, heading_text: "test")
     assert_select ".gem-c-image-card[data-module='gem-track-click ga4-link-tracker']"